requirements.txt
speechrecognition==3.8.1
pyttsx3==2.90
google-cloud-texttospeech==2.15.0
gtts==2.2.3
python-dotenv==1.0.0
google-generativeai==0.1.0
